Rename _npufunc to _npfunctions and get started.
