+++ tests/test_PR089_numpy_functions.py
@@ -1,0 +1,11 @@
+# BSD 3-Clause License; see https://github.com/jpivarski/awkward-1.0/blob/master/LICENSE
+
+import sys
+
+import pytest
+import numpy
+
+import awkward1
+
+def test_nothing():
+    pass
